requirements.txt
pygwarts
python-dotenv